[issue #23] fix clear patient capture data on edit
The data is now correctly cleared. This also affects the creation of new
patients where now the initial timestamp is set to null. However, to
ensure that every time entered is relative to the time of the incident
all timestamps are corrected now using the incident time as a reference.